refactor(tencent-mtg): extract sync logic to service and add operatorId support
- extract duplicate meeting/recording sync logic to dedicated mapper and service
- add operatorId parameter to sync recording APIs
- add time_zone field to meeting detail type definition
- create new record mapper utility file for shared conversion logic
- simplify processor by reusing the new sync service
